First commit of thredds-test-environment
Fully working packer + ansible configuration to provision an environment
for testing the various THREDDS related projects packaged as a Docker
Image and an AMI. AMI is currently used for worker nodes on our AWS
based Jenkins instance. The Docker Image is intended to be used as
the basis for a github action.